rename utility.h to utility.hpp
