+++ include/toolbox/container/tuple.hpp
@@ -1,7 +1,7 @@
 #ifndef _TOOLBOX_CONTAINER_TUPLE_HPP_
 #define _TOOLBOX_CONTAINER_TUPLE_HPP_
 
-#include <toolbox/std/type_traits.h>
+#include <toolbox/std/type_traits.hpp>
 #include <toolbox/std/utility.h>
 
 namespace toolbox
