+++ include/toolbox/std/utility.h
@@ -7,7 +7,7 @@
     #include <utility>
 #else
 
-#include "type_traits.h"
+#include "type_traits.hpp"
 
 namespace std
 {
